agents/a2a_agent_master/requirements.txt
uvicorn
google-adk==1.21.0
a2a-sdk==0.2.16
